requirements.txt
streamlit==1.7.0
SpeechRecognition==3.8.1
Pillow==8.3.2
openai==0.27.0
